Check EmailAddress for angles before parentheses
This fixes isse #218, though there are still edge cases for valid
emails which won't be allowed (such as those with comments) or
invalid emails which will still be allowed
(eg. this"is"invalid@email.com). The best choice for accuracy would
be to add a parser, but this could slow down runtime. It may not
be a worthy sacrifice.